requirements.txt
Flask
pytesseract
Pillow
transformers
torch
tf-keras